+++ mongoengine/fields.py
@@ -39,13 +39,13 @@ __all__ = [
     'FloatField', 'DecimalField', 'BooleanField', 'DateTimeField',
     'ComplexDateTimeField', 'EmbeddedDocumentField', 'ObjectIdField',
     'GenericEmbeddedDocumentField', 'DynamicField', 'ListField',
-    'SortedListField', 'DictField', 'MapField', 'ReferenceField',
-    'CachedReferenceField', 'GenericReferenceField', 'BinaryField',
-    'GridFSError', 'GridFSProxy', 'FileField', 'ImageGridFsProxy',
-    'ImproperlyConfigured', 'ImageField', 'GeoPointField', 'PointField',
-    'LineStringField', 'PolygonField', 'SequenceField', 'UUIDField',
-    'MultiPointField', 'MultiLineStringField', 'MultiPolygonField',
-    'GeoJsonBaseField']
+    'SortedListField', 'EmbeddedDocumentListField', 'DictField',
+    'MapField', 'ReferenceField', 'CachedReferenceField',
+    'GenericReferenceField', 'BinaryField', 'GridFSError', 'GridFSProxy',
+    'FileField', 'ImageGridFsProxy', 'ImproperlyConfigured', 'ImageField',
+    'GeoPointField', 'PointField', 'LineStringField', 'PolygonField',
+    'SequenceField', 'UUIDField', 'MultiPointField', 'MultiLineStringField',
+    'MultiPolygonField', 'GeoJsonBaseField']
 
 
 RECURSIVE_REFERENCE_CONSTANT = 'self'
@@ -726,6 +726,32 @@ class ListField(ComplexBaseField):
                 return [self.field.prepare_query_value(op, v) for v in value]
             return self.field.prepare_query_value(op, value)
         return super(ListField, self).prepare_query_value(op, value)
+
+
+class EmbeddedDocumentListField(ListField):
+    """A :class:`~mongoengine.ListField` designed specially to hold a list of
+    embedded documents to provide additional query helpers.
+
+    .. note::
+        The only valid list values are subclasses of
+        :class:`~mongoengine.EmbeddedDocument`.
+
+    .. versionadded:: 0.9
+
+    """
+
+    def __init__(self, document_type, *args, **kwargs):
+        """
+        :param document_type: The type of
+         :class:`~mongoengine.EmbeddedDocument` the list will hold.
+        :param args: Arguments passed directly into the parent
+         :class:`~mongoengine.ListField`.
+        :param kwargs: Keyword arguments passed directly into the parent
+         :class:`~mongoengine.ListField`.
+        """
+        super(EmbeddedDocumentListField, self).__init__(
+            field=EmbeddedDocumentField(document_type), **kwargs
+        )
 
 
 class SortedListField(ListField):